+++ tomcat/src/main/java/org/apache/coyote/http11/HttpStatus.java
@@ -6,6 +6,7 @@ public enum HttpStatus {
     BAD_REQUEST(400, "Bad Request"),
     UNAUTHORIZED(401, "Unauthorized"),
     NOT_FOUND(404, "Not Found"),
+    METHOD_NOT_ALLOWED(405, "Method Not Allowed"),
     INTERNAL_SERVER_ERROR(500, "Internal Server Error");
 
     private final int value;
