+++ tomcat/src/main/java/org/apache/coyote/http11/exception/HttpRequestMethodNotSupportedException.java
@@ -1,0 +1,10 @@
+package org.apache.coyote.http11.exception;
+
+import org.apache.coyote.http11.exception.base.BadRequestException;
+
+public class HttpRequestMethodNotSupportedException extends BadRequestException {
+
+    public HttpRequestMethodNotSupportedException() {
+        super(ErrorCode.METHOD_NOT_ALLOWED);
+    }
+}
